# core-common-document-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-document-mgmt
  description: Enterprise Content Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: https://api.getfirefly.io/ecm
    description: Production Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/documents/{id}:
    get:
      tags:
        - Document Controller
      summary: Get document by ID
      description: Returns a document by its ID
      operationId: getDocumentById
      parameters:
        - name: id
          in: path
          description: ID of the document to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved document
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
    put:
      tags:
        - Document Controller
      summary: Update an existing document
      description: Updates an existing document by its ID
      operationId: updateDocument
      parameters:
        - name: id
          in: path
          description: ID of the document to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentDTO'
        required: true
      responses:
        '200':
          description: Document updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '400':
          description: Invalid document data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
    delete:
      tags:
        - Document Controller
      summary: Delete a document
      description: Deletes a document by its ID
      operationId: deleteDocument
      parameters:
        - name: id
          in: path
          description: ID of the document to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Document deleted successfully
        '404':
          description: Document not found
...
  /api/v1/documents/{id}/versions:
    post:
      tags:
        - Document Controller
      summary: Create new document version
      description: Creates a new version of a document with new content
      operationId: createVersion
      parameters:
        - name: id
          in: path
          description: ID of the document to create version for
          required: true
          schema:
            type: string
            format: uuid
        - name: comment
          in: query
          description: Version comment
          required: false
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  description: File for new version
                  format: binary
      responses:
        '200':
          description: Document version created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '400':
          description: Invalid file upload
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
  /api/v1/documents/{id}/upload:
    post:
      tags:
        - Document Controller
      summary: Upload document content
      description: Uploads file content for an existing document
      operationId: uploadContent
      parameters:
        - name: id
          in: path
          description: ID of the document to upload content for
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  description: File to upload
                  format: binary
      responses:
        '200':
          description: Document content uploaded successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '400':
          description: Invalid file upload
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
...
  /api/v1/documents/{id}/metadata:
    get:
      tags:
        - Document Controller
      summary: Get document content metadata
      description: Retrieves metadata about document content
      operationId: getContentMetadata
      parameters:
        - name: id
          in: path
          description: ID of the document
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Document metadata retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
        '404':
          description: Document not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentDTO'
  /api/v1/documents/{id}/download:
    get:
      tags:
        - Document Controller
      summary: Download document content
      description: Downloads the content of a document
      operationId: downloadContent
      parameters:
        - name: id
          in: path
          description: ID of the document to download
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Party-ID
          in: header
          description: Identifier of the client (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Employee-ID
          in: header
          description: Identifier of the employee (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Service-Account-ID
          in: header
          description: Identifier of the service account (at least one of X-Party-ID, X-Employee-ID, or X-Service-Account-ID is required)
          required: false
          schema:
            type: string
        - name: X-Auth-Roles
          in: header
          description: Roles of the subject (CUSTOMER, ADMIN, CUSTOMER_SUPPORT, SUPERVISOR, MANAGER, BRANCH_STAFF, SERVICE_ACCOUNT), comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Auth-Scopes
          in: header
          description: OAuth2 scopes like contracts.read, accounts.write, comma-separated (optional)
          required: false
          schema:
            type: string
        - name: X-Request-ID
          in: header
          description: Unique identifier for the request, used for traceability (optional)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Document content downloaded successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataBuffer'
        '404':
          description: Document content not available
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataBuffer'
...
components:
  schemas:
...
    DocumentDTO:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the document
          format: uuid
          readOnly: true
        name:
          type: string
          description: Name of the document
        description:
          type: string
          description: Description of the document
        fileName:
          type: string
          description: Original file name of the document
        fileExtension:
          type: string
          description: File extension of the document
        mimeType:
          type: string
          description: MIME type of the document
        fileSize:
          type: integer
          description: Size of the document file in bytes
          format: int64
        documentType:
          type: string
          description: Type of the document
          enum:
            - DOCUMENT
            - IMAGE
            - PDF
            - CONTRACT
            - INVOICE
            - RECEIPT
            - FORM
            - REPORT
            - EMAIL
            - AUDIO
            - VIDEO
            - ARCHIVE
            - OTHER
        documentStatus:
          type: string
          description: Current status of the document
          enum:
            - DRAFT
            - UNDER_REVIEW
            - APPROVED
            - REJECTED
            - PUBLISHED
            - ARCHIVED
            - MARKED_FOR_DELETION
            - DELETED
            - LOCKED
            - EXPIRED
        storageType:
          type: string
          description: Storage type of the document
          enum:
            - LOCAL_FILESYSTEM
            - DATABASE
            - S3
            - AZURE_BLOB
            - GOOGLE_CLOUD_STORAGE
            - CDN
            - DISTRIBUTED_FS
            - EXTERNAL_REFERENCE
        storagePath:
          type: string
          description: Path where the document is stored
        securityLevel:
          type: string
          description: Security level of the document
          enum:
            - PUBLIC
            - INTERNAL
            - CONFIDENTIAL
            - RESTRICTED
            - SECRET
            - TOP_SECRET
        folderId:
          type: string
          description: ID of the folder containing the document
          format: uuid
        isEncrypted:
          type: boolean
          description: Indicates if the document is encrypted
        isIndexed:
          type: boolean
          description: Indicates if the document is indexed for search
        isLocked:
          type: boolean
          description: Indicates if the document is locked for editing
        lockedBy:
          type: string
          description: User who locked the document
        lockedUntil:
          type: string
          description: Date and time until the document is locked
          format: date-time
        expirationDate:
          type: string
          description: Date and time when the document expires
          format: date-time
        retentionDate:
          type: string
          description: Date and time until the document should be retained
          format: date-time
        tenantId:
          type: string
          description: Tenant ID for multi-tenancy support
        createdAt:
          type: string
          description: Date and time when the document was created
          format: date-time
        createdBy:
          type: string
          description: User who created the document
        updatedAt:
          type: string
          description: Date and time when the document was last updated
          format: date-time
        updatedBy:
          type: string
          description: User who last updated the document
        version:
          type: integer
          description: Version number for optimistic locking
          format: int64
        checksum:
          type: string
          description: Checksum of the document file for integrity verification
      description: Document data transfer object
...
    DataBuffer:
      type: object
...